Rename requirements composition module
Renames the internal requirements_composition module to requirements_layers so the module name matches the layer-oriented model used by the PR. The public compose_requirements API remains unchanged; this is only an internal organization/name cleanup plus matching wording for composite requirement sources and hook comments.
