+++ codex-rs/config/src/lib.rs
@@ -17,8 +17,8 @@ pub mod permissions_toml;
 mod plugin_edit;
 pub mod profile_toml;
 mod project_root_markers;
-mod requirements_composition;
 mod requirements_exec_policy;
+mod requirements_layers;
 pub mod schema;
 mod skills_config;
 mod state;
@@ -112,14 +112,14 @@ pub use plugin_edit::clear_user_plugin;
 pub use plugin_edit::set_user_plugin_enabled;
 pub use project_root_markers::default_project_root_markers;
 pub use project_root_markers::project_root_markers_from_config;
-pub use requirements_composition::RequirementsLayerEntry;
-pub use requirements_composition::compose_requirements;
 pub use requirements_exec_policy::RequirementsExecPolicy;
 pub use requirements_exec_policy::RequirementsExecPolicyDecisionToml;
 pub use requirements_exec_policy::RequirementsExecPolicyParseError;
 pub use requirements_exec_policy::RequirementsExecPolicyPatternTokenToml;
 pub use requirements_exec_policy::RequirementsExecPolicyPrefixRuleToml;
 pub use requirements_exec_policy::RequirementsExecPolicyToml;
+pub use requirements_layers::RequirementsLayerEntry;
+pub use requirements_layers::compose_requirements;
 pub use skills_config::BundledSkillsConfig;
 pub use skills_config::SkillConfig;
 pub use skills_config::SkillsConfig;
